// sdks/node-ts/src/apache_beam/proto/beam_runner_api.ts
export interface Components {
    /**
     * (Required) A map from pipeline-scoped id to PTransform.
     *
     * Keys of the transforms map may be used by runners to identify pipeline
     * steps. Hence it's recommended to use strings that are not too long that
     * match regex '[A-Za-z0-9-_]+'.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PTransform> transforms = 1;
     */
    transforms: {
        [key: string]: PTransform;
    };
    /**
     * (Required) A map from pipeline-scoped id to PCollection.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PCollection> pcollections = 2;
     */
    pcollections: {
        [key: string]: PCollection;
    };
    /**
     * (Required) A map from pipeline-scoped id to WindowingStrategy.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.WindowingStrategy> windowing_strategies = 3;
     */
    windowingStrategies: {
        [key: string]: WindowingStrategy;
    };
    /**
     * (Required) A map from pipeline-scoped id to Coder.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Coder> coders = 4;
     */
    coders: {
        [key: string]: Coder;
    };
    /**
     * (Required) A map from pipeline-scoped id to Environment.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Environment> environments = 5;
     */
    environments: {
        [key: string]: Environment;
    };
}
// ...
export interface Pipeline {
    /**
     * (Required) The coders, UDFs, graph nodes, etc, that make up
     * this pipeline.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.Components components = 1;
     */
    components?: Components;
    /**
     * (Required) The ids of all PTransforms that are not contained within another
     * PTransform. These must be in shallow topological order, so that traversing
     * them recursively in this order yields a recursively topological traversal.
     *
     * @generated from protobuf field: repeated string root_transform_ids = 2;
     */
    rootTransformIds: string[];
    /**
     * (Optional) Static display data for the pipeline. If there is none,
     * it may be omitted.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.DisplayData display_data = 3;
     */
    displayData: DisplayData[];
    /**
     * (Optional) A set of requirements that the runner MUST understand and be
     * able to faithfully provide in order to execute this pipeline. These
     * may indicate that a runner must inspect new fields on a component or
     * provide additional guarantees when processing specific transforms.
     * A runner should reject any pipelines with unknown requirements.
     *
     * @generated from protobuf field: repeated string requirements = 4;
     */
    requirements: string[];
}
// ...
export interface PTransform {
    /**
     * (Required) A unique name for the application node.
     *
     * Ideally, this should be stable over multiple evolutions of a pipeline
     * for the purposes of logging and associating pipeline state with a node,
     * etc.
     *
     * If it is not stable, then the runner decides what will happen. But, most
     * importantly, it must always be here and be unique, even if it is
     * autogenerated.
     *
     * @generated from protobuf field: string unique_name = 5;
     */
    uniqueName: string;
    /**
     * (Optional) A URN and payload that, together, fully defined the semantics
     * of this transform.
     *
     * If absent, this must be an "anonymous" composite transform.
     *
     * For primitive transform in the Runner API, this is required, and the
     * payloads are well-defined messages. When the URN indicates ParDo it
     * is a ParDoPayload, and so on. For some special composite transforms,
     * the payload is also officially defined. See StandardPTransforms for
     * details.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FunctionSpec spec = 1;
     */
    spec?: FunctionSpec;
    /**
     * (Optional) A list of the ids of transforms that it contains.
     *
     * Primitive transforms (see StandardPTransforms.Primitives) are not allowed
     * to specify subtransforms.
     *
     * Note that a composite transform may have zero subtransforms as long as it
     * only outputs PCollections that are in its inputs.
     *
     * @generated from protobuf field: repeated string subtransforms = 2;
     */
    subtransforms: string[];
    /**
     * (Required) A map from local names of inputs (unique only with this map, and
     * likely embedded in the transform payload and serialized user code) to
     * PCollection ids.
     *
     * The payload for this transform may clarify the relationship of these
     * inputs. For example:
     *
     *  - for a Flatten transform they are merged
     *  - for a ParDo transform, some may be side inputs
     *
     * All inputs are recorded here so that the topological ordering of
     * the graph is consistent whether or not the payload is understood.
     *
     * @generated from protobuf field: map<string, string> inputs = 3;
     */
    inputs: {
        [key: string]: string;
    };
    /**
     * (Required) A map from local names of outputs (unique only within this map,
     * and likely embedded in the transform payload and serialized user code)
     * to PCollection ids.
     *
     * The URN or payload for this transform node may clarify the type and
     * relationship of these outputs. For example:
     *
     *  - for a ParDo transform, these are tags on PCollections, which will be
     *    embedded in the DoFn.
     *
     * @generated from protobuf field: map<string, string> outputs = 4;
     */
    outputs: {
        [key: string]: string;
    };
    /**
     * (Optional) Static display data for this PTransform application. If
     * there is none, it may be omitted.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.DisplayData display_data = 6;
     */
    displayData: DisplayData[];
    /**
     * Environment where the current PTransform should be executed in.
     *
     * Transforms that are required to be implemented by a runner must omit this.
     * All other transforms are required to specify this.
     *
     * @generated from protobuf field: string environment_id = 7;
     */
    environmentId: string;
    /**
     * (Optional) A map from URNs designating a type of annotation, to the
     * annotation in binary format. For example, an annotation could indicate
     * that this PTransform has specific privacy properties.
     *
     * A runner MAY ignore types of annotations it doesn't understand. Therefore
     * annotations MUST NOT be used for metadata that can affect correct
     * execution of the transform.
     *
     * @generated from protobuf field: map<string, bytes> annotations = 8;
     */
    annotations: {
        [key: string]: Uint8Array;
    };
}
// ...
export interface DisplayData {
    /**
     * A key used to describe the type of display data. See StandardDisplayData
     * for the set of well known urns describing how the payload is meant to be
     * interpreted.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Optional) The data specifying any parameters to the URN. If
     * the URN does not require any arguments, this may be omitted.
     *
     * @generated from protobuf field: bytes payload = 2;
     */
    payload: Uint8Array;
}
// ...
export const Components = new Components$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Pipeline$Type extends MessageType<Pipeline> {
    constructor() {
        super("org.apache.beam.model.pipeline.v1.Pipeline", [
            { no: 1, name: "components", kind: "message", T: () => Components },
            { no: 2, name: "root_transform_ids", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "display_data", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => DisplayData },
            { no: 4, name: "requirements", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<Pipeline>): Pipeline {
        const message = { rootTransformIds: [], displayData: [], requirements: [] };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Pipeline>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Pipeline): Pipeline {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* org.apache.beam.model.pipeline.v1.Components components */ 1:
                    message.components = Components.internalBinaryRead(reader, reader.uint32(), options, message.components);
                    break;
                case /* repeated string root_transform_ids */ 2:
                    message.rootTransformIds.push(reader.string());
                    break;
                case /* repeated org.apache.beam.model.pipeline.v1.DisplayData display_data */ 3:
                    message.displayData.push(DisplayData.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated string requirements */ 4:
                    message.requirements.push(reader.string());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Pipeline, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* org.apache.beam.model.pipeline.v1.Components components = 1; */
        if (message.components)
            Components.internalBinaryWrite(message.components, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated string root_transform_ids = 2; */
        for (let i = 0; i < message.rootTransformIds.length; i++)
            writer.tag(2, WireType.LengthDelimited).string(message.rootTransformIds[i]);
        /* repeated org.apache.beam.model.pipeline.v1.DisplayData display_data = 3; */
        for (let i = 0; i < message.displayData.length; i++)
            DisplayData.internalBinaryWrite(message.displayData[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* repeated string requirements = 4; */
        for (let i = 0; i < message.requirements.length; i++)
            writer.tag(4, WireType.LengthDelimited).string(message.requirements[i]);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.pipeline.v1.Pipeline
 */
export const Pipeline = new Pipeline$Type();
// ...
export const PTransform = new PTransform$Type();
// ...
export const PCollection = new PCollection$Type();
// ...
export const Coder = new Coder$Type();
// ...
export const WindowingStrategy = new WindowingStrategy$Type();
// ...
export const Environment = new Environment$Type();
// ...
export const DisplayData = new DisplayData$Type();