// sdks/node-ts/src/apache_beam/proto/beam_job_api.ts
export enum JobState_Enum {
    /**
     * The job state reported by a runner cannot be interpreted by the SDK.
     *
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * The job has not yet started.
     *
     * @generated from protobuf enum value: STOPPED = 1;
     */
    STOPPED = 1,
    /**
     * The job is currently running.
     *
     * @generated from protobuf enum value: RUNNING = 2;
     */
    RUNNING = 2,
    /**
     * The job has successfully completed. (terminal)
     *
     * @generated from protobuf enum value: DONE = 3;
     */
    DONE = 3,
    /**
     * The job has failed. (terminal)
     *
     * @generated from protobuf enum value: FAILED = 4;
     */
    FAILED = 4,
    /**
     * The job has been explicitly cancelled. (terminal)
     *
     * @generated from protobuf enum value: CANCELLED = 5;
     */
    CANCELLED = 5,
    /**
     * The job has been updated. (terminal)
     *
     * @generated from protobuf enum value: UPDATED = 6;
     */
    UPDATED = 6,
    /**
     * The job is draining its data. (optional)
     *
     * @generated from protobuf enum value: DRAINING = 7;
     */
    DRAINING = 7,
    /**
     * The job has completed draining its data. (terminal)
     *
     * @generated from protobuf enum value: DRAINED = 8;
     */
    DRAINED = 8,
    /**
     * The job is starting up.
     *
     * @generated from protobuf enum value: STARTING = 9;
     */
    STARTING = 9,
    /**
     * The job is cancelling. (optional)
     *
     * @generated from protobuf enum value: CANCELLING = 10;
     */
    CANCELLING = 10,
    /**
     * The job is in the process of being updated. (optional)
     *
     * @generated from protobuf enum value: UPDATING = 11;
     */
    UPDATING = 11
}